moved trio binary to main folder
