+++ src/projects/lja/CMakeLists.txt
@@ -9,6 +9,9 @@ target_link_libraries(lja lja_pipeline)
 add_executable(jumboDBG dbg.cpp subdataset_processing.cpp)
 target_link_libraries(jumboDBG lja_ec lja_dbg lja_common lja_sequence m)
 
+add_executable(run_trio run_trio.cpp)
+target_link_libraries(run_trio lja_trio)
+
 install(PROGRAMS "${PYTHON_DIR}/run_rr.py"
         DESTINATION bin
         COMPONENT runtime)
