+++ src/projects/lja/run_trio.cpp
@@ -1,10 +1,11 @@
 #include <common/cl_parser.hpp>
-#include "trio.hpp"
+#include "trio/trio.hpp"
+#include "pipeline.hpp"
 
 using namespace trio;
 
 int main(int argc, char **argv) {
-    CLParser parser({"diplo_graph=", "haployak=", "output=", "debug=none", "threads=8", "corrected_reads=", "bridge_cutoff=1000000"}, {"reads"},
+    CLParser parser({"diplo_graph=", "haployak=", "output=", "debug=none", "threads=8", "corrected_reads=", "bridge_cutoff=1000000"}, {"reads", "ref"},
                     {});
 
     parser.parseCL(argc, argv);
@@ -14,6 +15,8 @@ int main(int argc, char **argv) {
         std::cout << parser.check() << std::endl;
         return 1;
     }
+
+    pipeline::LJAPipeline aux_pipeline(oneline::initialize<std::experimental::filesystem::path>(parser.getListValue("ref")));
     std::experimental::filesystem::path dir(parser.getValue("output"));
     ensure_dir_existance(dir);
     logging::LoggerStorage ls(dir, "dbg");
@@ -33,9 +36,9 @@ int main(int argc, char **argv) {
     io::Library reads_lib = oneline::initialize<std::experimental::filesystem::path>(parser.getListValue("reads"));
 
     std::experimental::filesystem::path res_m(dir / "graph_p.gfa");
-    simplifyHaplo(logger, threads, res_m, graph, haplo, 'm', corrected_reads, reads_lib, dir, saved_bridge_cutoff);
+    aux_pipeline.simplifyHaplo(logger, threads, res_m, graph, haplo, 'm', corrected_reads, reads_lib, dir, saved_bridge_cutoff);
     std::experimental::filesystem::path res_p(dir / "graph_m.gfa");
-    simplifyHaplo(logger, threads, res_p, graph, haplo, 'p', corrected_reads, reads_lib, dir, saved_bridge_cutoff);
+    aux_pipeline.simplifyHaplo(logger, threads, res_p, graph, haplo, 'p', corrected_reads, reads_lib, dir, saved_bridge_cutoff);
 
 
 }
