+++ src/projects/trio/trio.hpp
@@ -46,14 +46,4 @@ struct HaplotypeRemover {
     void removeHaplotype();
 };
 
-std::experimental::filesystem::path simplifyHaplo(logging::Logger &logger, size_t threads,
-                                                  const std::experimental::filesystem::path &output_file,
-                                                  const std::experimental::filesystem::path &diplo_graph,
-                                                  const std::experimental::filesystem::path &haployak,
-                                                  const char haplotype,
-                                                  const std::experimental::filesystem::path &corrected_reads,
-                                                  const io::Library &reads,
-                                                  const std::experimental::filesystem::path &dir,
-                                                  const size_t saved_bridge_cutoff);
-
 }
